test(api): share FakeClock across admission unit tests
Deduplicate the manually-advanced FakeClock helper into a local
tests/unit/api/admission/helpers.py module and import it in both
test_slot_pool.py and test_codel.py.
